requirements.txt
# ===============================
# CORE GOOGLE CLOUD / AI
# ===============================
google-cloud-storage==2.18.2
google-cloud-aiplatform==1.70.0
google-generativeai==0.7.2
google-api-python-client==2.154.0
google-auth==2.35.0
google-auth-oauthlib==1.2.1
google-auth-httplib2==0.2.0
google-cloud-speech==2.33.0

# ===============================
# API / SERVER
# ===============================
fastapi==0.115.2
uvicorn==0.31.1
pydantic==2.9.2
tenacity==9.0.0

# ===============================
# AUDIO PROCESSING
# ===============================
pydub==0.25.1
ffmpeg-python==0.2.0

# ===============================
# DOCUMENT / REPORTING
# ===============================
reportlab==4.2.5
PyPDF2==3.0.1
python-docx==1.1.2
PyMuPDF==1.24.10
Jinja2==3.1.4

# ===============================
# DATA / AUTOMATION
# ===============================
gspread==6.1.4
python-dotenv==1.0.1
polars==1.12.0
